+++ log.go
@@ -1,12 +1,13 @@
 package main
 
 import (
+	"fmt"
 	"net/http"
 
 	logrus "github.com/sirupsen/logrus"
 )
 
-func initLog() {
+func initLog() error {
 	logFormat := "pretty"
 	strEnvConfig(&logFormat, "IMGPROXY_LOG_FORMAT")
 
@@ -32,11 +33,13 @@ func initLog() {
 	if isSyslogEnabled() {
 		slHook, err := newSyslogHook()
 		if err != nil {
-			logFatal("Unable to connect to local syslog daemon")
+			return fmt.Errorf("Unable to connect to syslog daemon: %s", err)
 		}
 
 		logrus.AddHook(slHook)
 	}
+
+	return nil
 }
 
 func logRequest(reqID string, r *http.Request) {
@@ -94,6 +97,10 @@ func logNotice(f string, args ...interface{}) {
 
 func logWarning(f string, args ...interface{}) {
 	logrus.Warnf(f, args...)
+}
+
+func logError(f string, args ...interface{}) {
+	logrus.Errorf(f, args...)
 }
 
 func logFatal(f string, args ...interface{}) {